XDP Offload: Handle redirects to different interfaces [skip ci]
The idea: set the source IP address of the outgoind packet to the
sender NIC's address. This enables relaying traffic on arbitrary
interfaces.

Caveat: works on outgoing traffic only, reverse traffic (e.g., from
the peer on a diff NIC) will not be XDP-accelerated because we use the
RelayAddress in channel bindings and that is not the NIC IP address..