XDP offload: Fix downstream in multi NIC setup [skip ci]
The local IP addr is not known apriori if the server listens on
multiple interfaces. The peer connection uses the server's public
address when upserted into the offload. However, the peer might
connect via a different NIC thus having a local IP address that is not
the public IP address. To overcome this issue, insert peer connection
with local IP set to 0 to mask it during matching.